Move some code arround
